GH-51 - moved audio stuff for preview to separate component
